requirements_free_tier.txt
# FREE TIER Optimized Requirements (Memory Efficient)
# Core FastAPI and web dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
python-multipart==0.0.6

# Data handling and validation
pydantic==2.5.0
python-dateutil==2.8.2

# Lightweight ML libraries for FREE TIER
torch==2.1.1 --index-url https://download.pytorch.org/whl/cpu  # CPU-only version
torchvision==0.16.1 --index-url https://download.pytorch.org/whl/cpu
transformers==4.36.0
clip-by-openai==1.0
pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2

# Essential image processing
opencv-python-headless==4.8.1.78  # Headless version (smaller)
requests==2.31.0

# Lightweight text processing
nltk==3.8.1
regex==2023.10.3

# Web scraping (minimal)
beautifulsoup4==4.12.2
lxml==4.9.3

# FREE TIER: Essential metadata analysis only
exifread==3.0.0
python-magic==0.4.27

# FREE TIER: Lightweight caching
redis==5.0.1

# FREE TIER: Smaller models
# distilbert-base-uncased (smaller than bert-base-uncased)
# sentence-transformers==2.2.2  # Skip for now to save memory

# Performance monitoring (lightweight)
psutil==5.9.6

# Cloud integration (minimal)
boto3==1.34.0
python-dotenv==1.0.0

# Development and testing (minimal)
pytest==7.4.3
pytest-asyncio==0.21.1

# Text utilities
ftfy==6.1.1